requirements.txt
asgiref==3.6.0
coverage==7.2.7
Django==4.2
django-extensions==3.2.3
djangorestframework==3.14.0
numpy==1.25.0
pandas==2.0.3
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
